update components struture
