--- src/components/users/BaseTextField.vue
@@ -1,9 +1,0 @@
-<template>
-  <v-text-field v-model="model" outlined color="primary"></v-text-field>
-</template>
-
-<script setup lang="ts">
-const model = defineModel();
-</script>
-
-<style scoped></style>
